Move QueryFhirPlugin to ChatApp.Server
