+++ src/ChatApp/ChatApp.Server/Program.cs
@@ -1,3 +1,5 @@
+using ChatApp.Server.Models;
+
 namespace ChatApp.Server;
 
 public class Program
@@ -8,6 +10,7 @@ public class Program
 
         builder.Services.Configure<CosmosOptions>(builder.Configuration.GetSection(nameof(CosmosOptions)));
         builder.Services.Configure<StorageOptions>(builder.Configuration.GetSection(nameof(StorageOptions)));
+        builder.Services.Configure<FhirOptions>(builder.Configuration.GetSection(nameof(FhirOptions)));
 
         // Add services to the container.
         builder.Services.AddAuthorization();
